requirements.txt
django>=5.0,<6.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-filter>=24.0
drf-spectacular>=0.27
python-dotenv>=1.0
gunicorn>=21.0